requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==6.0.2
django-crispy-forms==2.6
django-filter==25.2
djangorestframework==3.17.0
Markdown==3.10.2
sqlparse==0.5.5
